requirements.txt
asgiref==3.6.0
black==23.3.0
click==8.1.3
crispy-bootstrap5==0.7
Django==4.2
django-crispy-forms==2.0
django-maintenance-mode==0.18.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.1
psycopg==3.1.9
python-fsutil==0.10.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.5.0
whitenoise==6.5.0
